Actualizado el controlador de usuario para mayor seguridad.

+++ src/user/user.controller.ts
@@ -8,11 +8,22 @@ import {
   Param,
   Patch,
   Post,
+  UseGuards,
+  Request,
+  ForbiddenException,
+  ParseUUIDPipe,
 } from '@nestjs/common';
-import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import {
+  ApiBearerAuth,
+  ApiOperation,
+  ApiResponse,
+  ApiTags,
+} from '@nestjs/swagger';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import type { AuthenticatedRequest } from '../common/interfaces/authenticated-request.interface';
 
 @ApiTags('user')
 @Controller('user')
@@ -39,15 +50,34 @@ export class UserController {
   }
 
   @Patch(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Partially update user' })
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  @ApiResponse({ status: 200, description: 'User updated successfully' })
+  @ApiResponse({
+    status: 403,
+    description: 'You do not have the permission to modify this user',
+  })
+  async update(
+    @Param('id', ParseUUIDPipe) id: string,
+    @Body() updateUserDto: UpdateUserDto,
+    @Request() req: AuthenticatedRequest,
+  ) {
+    if (req.user.userId != id) {
+      throw new ForbiddenException('You cannot modify the data of other user');
+    }
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @HttpCode(HttpStatus.NO_CONTENT)
   @ApiOperation({ summary: 'Delete an user' })
-  async remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @Request() req: AuthenticatedRequest) {
+    if (req.user.userId != id) {
+      throw new ForbiddenException('You cannot delete other user account');
+    }
     await this.usersService.remove(id);
   }
 }
